+++ Changes.md
@@ -4,6 +4,10 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
+## HEAD
+
+* Delay implied connect until after `.query` argument validation
+
 ## v2.6.2 (2015-04-14)
 
 * Fix `Connection.createQuery` for no SQL #1058