+++ Changes.md
@@ -6,6 +6,7 @@ you spot any mistakes.
 
 ## HEAD
 
+* Destroy/end connections removed from the pool on error
 * Delay implied connect until after `.query` argument validation
 * Do not remove connections with non-fatal errors from the pool
 * Error early if `callback` argument to `.query` is not a function #1060