+++ Changes.md
@@ -7,6 +7,7 @@ you spot any mistakes.
 ## NEXT
 
 * Add code `POOL_NOEXIST` in PoolCluster error #846
+* Default `connectTimeout` to 10 seconds
 * Return `Query` object from `pool.query` like `conn.query` #830
 * Use `EventEmitter.listenerCount` when possible for faster counting